test(StatusCode): adds a test for investigating a potential bug
Issues #255 indicates that there could be a bug with the statusCodeIsNotXXX methods. The assertion log Messages showed "UNKNOWN_ACTUAL_VALUE_forStatusCodeAssertion" in the actual field.

The bug could not be reproduced. It is likely that it was already been fix.

But the corresponding test case has been added.

____
Reference: Issues #255