fix(ApiTestExecutor): fix wrong Status and Message in Logs
There was a bug which only occurred in very special cases.
This caused the status of the ApiAssertionLogMessages to be set
incorrectly and the logged error message did not match the actual
error message.

The bug occurred when several assertions were checked in an Api
test and the first assertion failed, but the second should be
successful and the assertions were of different types (for example TimeAssertion and StatusCodeAssertion).

This is because restAssured checks the current and
previous assertions for each ValidatableResponse.
The desired behaviour for QTAF could be implemented by
appending only one assertion to a new ValidatableResponse
in the ApiTestExecutor.