fix(ApiLogMessage): logs get updated with the values of the actual request
Before that the values a user has set got logged.
But it was not guarantied that RESTassured
just sends the values set by the user as a request.
Sometimes RESTassured internally adds
for example headers that are implicitly needed.
That's way in the implementation before this
did not represent the actual request
that RESTassure sends correctly in the QTAF-logs