refactor(ApiRequestSpecification): remove logMessage parameters of functional interface
The log message parameter was an artefact of the previous system
for updating the logMessage. It was needed to change the log Message
depending on which request specification method got called.
Some commits ago there was introduced a new approach
for updating the Request of the logMessage
in a single place in the ApiExecutor.
That's why this attribute is not needed anymore.

The corresponding "apply"-method needed to be updated
in the ApiTestExecutor.
In addition some variables in the ApiTestExecutor
got renamed for better readability.