feat(qtaf-logging)!: improve console log messages
The user can now configure if he wants the steps of a test case
to get logged. For shorter log messages,
the steps can be disabled in the qtaf.json.
In addition, the corresponding method of the test case is now
displayed for quicker debugging when analyzing the test results.

BREAKING: The hashCode won't get printed to the console anymore,
to shorten the message.