docs(PR-Template): improve check for updating docu
There was a conflict with the check for updating the official documentation. If the documentation gets updated before the code, then the documentation is ahead of the actual code. 
If the documentation gets updated after the code, then it is not possible to confirm at this point in time that the documentation has already been updated.