doc(.github): add pr template (#325)
* doc(.github): add pr template

This adds a template for pull requests. 
It is meant to be to standardize the process of pull requests and to
 help the author and the reviewer during this process.

* doc(pull_request_template.md): link to commit conventions

So that the reviewer can easily refer and check if the conventions of the commit messages are met, a link to the conventions is added.

* docs(PR-teamplate): shorten sentence

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>

* docs(PR-Template): remove checks for description

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>

* docs(PR-Template): add a more detailed desciption

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>

* docs(PR-Template): remove redundant comment

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>

* docs(PR-Template): simplify wording

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>

* docs(PR-Template): remove unnecessary comment

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>

* docs(PR-Temaple): fix typo

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>

* docs(PR-Temaple): fix typo

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>

* docs(PR-Temaple): fix typo

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>

* docs(PR-Template): improve check for updating docu

There was a conflict with the check for updating the official documentation. If the documentation gets updated before the code, then the documentation is ahead of the actual code. 
If the documentation gets updated after the code, then it is not possible to confirm at this point in time that the documentation has already been updated.

* docs(PR-Template): correct dir

* docs(PR-Temp): Delete .github/.github directory

---------

Co-authored-by: Basti <18561435+csvtuda@users.noreply.github.com>