TestNG Listener refactor (#328)
* fix(#311): add EventListener for TestNG Invocation

The status of a test case should not be changed
after termination. For this case TestNG provides
 an InvocationEventListener.

* docs(TestNG Listeners): add JavaDocs

* test(Xray): ignore test

Some tests failed because they depended on a Jira project
that no longer exists. Since the feature that was covered
by the test is not actively supported and mocking the
response did not lead to meaningful tests, the tests
were deactivated.

* test(InvocationListener): add tests

Test for the method of the TestNGInvocationEventListener have been added.
Therefore, it was necessary to manipulate the log index
during the tests because, depending on the failed
steps, the method under test behaves differently