fix(#327): remove hard coded headless in chrome remote
The method enableDownloadOnRemoteChromeDriver()
has set the mode to headless in the Driver Options.
This caused all tests performed with the Remote Chrome Driver
to be started in a headless browser,
regardless of which driver options were set in qtaf.json.
This behavior was not as desired.
This commit removes the hard coded set of headless.

This bug was introduced with version 0.2.20.
The version which contains this commit closes issue #327